+++ src/v8unpack/MetaDataObject/Form/FormElements4/Field.py
@@ -1,4 +1,5 @@
-from .FormElement import FormElement, calc_offset
+from .FormElement import FormElement
+from v8unpack.helper import calc_offset
 
 
 class Field(FormElement):
