+++ src/v8unpack/MetaDataObject/Form/FormElements4/FormElement.py
@@ -2,25 +2,7 @@ from enum import Enum
 
 from v8unpack import helper
 from v8unpack.ext_exception import ExtException
-
-
-def calc_offset(counters, raw_data):
-    # counters - позиции указывающие на счетчики, если не 0 то за ним идет столько записей размера size
-    #  [(3, 1), (1, 0)] (смещение относительно предыдущей записи, количество записей в единице)
-    index = 0
-    for counter_index, size in counters:
-        index += counter_index
-        if size:
-            try:
-                value = int(raw_data[index])
-            except Exception as err:
-                raise ExtException(
-                    message='bad offset',
-                    detail=f'{counter_index}={index}',
-                    dump={'counters': counters, 'value': raw_data[index]}
-                )
-            index += value * size
-    return index
+from v8unpack.helper import calc_offset
 
 
 def check_count_element(counters, raw_data):
