+++ src/v8unpack/version.py
@@ -1,2 +1,2 @@
-__version__ = '1.0.8'
+__version__ = '1.0.10'
